flaskapp/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.6
Flask-HTTPAuth==3.2.4
google-api-python-client==1.7.3
google-cloud-datastore==1.6.0
google-cloud-bigquery==1.3.0
gunicorn==19.8.1
passlib==1.7.1
jsonschema==2.6.0
